Authorization for both admin and user role successfully. Person who has user role can not be allowed to Create/Get/Update/Delete user's information
