+++ src/main/java/com/example/crm/controller/AuthenticationController.java
@@ -12,6 +12,7 @@ import com.example.crm.repository.UserRepository;
 import com.example.crm.security.jwt.JwtUtils;
 import com.example.crm.security.services.UserDetailsImpl;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,71 +39,7 @@ public class AuthenticationController {
     @Autowired
     AuthenticationManager authenticationManager;
     @Autowired
-    UserRepository userRepository;
-    @Autowired
-    RoleRepository roleRepository;
-    @Autowired
-    PasswordEncoder encoder;
-    @Autowired
     JwtUtils jwtUtils;
-
-    @Operation(summary = "Create a new user",
-                description = "Save information of a new user with: firstname, lastname, email, username, phone number, password, birthday, gender and role.")
-    @PostMapping("/signup")
-    public ResponseEntity<?> registerUser(@Valid @RequestBody SignupRequest signUpRequest) {
-        if (userRepository.existsByUsername(signUpRequest.getUsername())) {
-            return ResponseEntity
-                    .badRequest()
-                    .body(new MessageResponse("Error: Username is already taken!"));
-        }
-
-        if (userRepository.existsByEmail(signUpRequest.getEmail())) {
-            return ResponseEntity
-                    .badRequest()
-                    .body(new MessageResponse("Error: Email is already taken!"));
-        }
-
-        // Create new user's account
-        User user = new User(signUpRequest.getFirstname(),
-                signUpRequest.getLastname(),
-                 signUpRequest.getLastname() + " " +  signUpRequest.getFirstname(),
-                signUpRequest.getEmail(),
-                signUpRequest.getUsername(),
-                signUpRequest.getPhoneNumber(),
-                encoder.encode(signUpRequest.getPassword()),
-                signUpRequest.getBirthday(),
-                signUpRequest.getGender()
-        );
-
-        Set<String> strRoles = signUpRequest.getRole();
-        Set<Role> roles = new HashSet<>();
-
-        if (strRoles == null) {
-            Role userRole = roleRepository.findByName(ERole.ROLE_USER)
-                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-            roles.add(userRole);
-        } else {
-            strRoles.forEach(role -> {
-                switch (role) {
-                    case "admin":
-                        Role adminRole = roleRepository.findByName(ERole.ROLE_ADMIN)
-                                .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-                        roles.add(adminRole);
-
-                        break;
-                    default:
-                        Role userRole = roleRepository.findByName(ERole.ROLE_USER)
-                                .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-                        roles.add(userRole);
-                }
-            });
-        }
-
-        user.setRoles(roles);
-        userRepository.save(user);
-
-        return ResponseEntity.ok(new MessageResponse("User registered successfully!"));
-    }
 
     @Operation(summary = "Log in")
     @PostMapping("/signin")
