+++ src/main/java/com/example/crm/security/WebSecurityConfig.java
@@ -51,10 +51,10 @@ public class WebSecurityConfig {
                 .exceptionHandling(exception -> exception.authenticationEntryPoint(unauthorizedHandler))
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .authorizeHttpRequests(auth ->
-                        auth.requestMatchers("/auth/**").permitAll()
+                        auth.requestMatchers("/auth/signin").permitAll()
                                 .requestMatchers("/v3/api-docs/**", "/swagger-ui/**").permitAll()
-                                .requestMatchers("/user/**").hasAuthority("ADMIN")
-                                .requestMatchers("/customer/**").hasAuthority("USER")
+                                .requestMatchers("/user/**").hasAnyAuthority("ROLE_ADMIN")
+                                .requestMatchers("/customer/**").hasAnyAuthority("ROLE_USER", "ROLE_ADMIN")
                         .anyRequest().authenticated()
                 );
         http.authenticationProvider(authenticationProvider());
